// backend/src/lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  OWNER: 'OWNER',
  MEMBER: 'MEMBER',
  ADMIN: 'ADMIN',
  ASSISTANT: 'ASSISTANT',
  VIEWER: 'VIEWER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const DocumentType = {
  RUNBOOK: 'RUNBOOK',
  ARCHITECTURE_SPEC: 'ARCHITECTURE_SPEC',
  PAST_INCIDENT_RCA: 'PAST_INCIDENT_RCA',
  API_DOC: 'API_DOC'
} as const

export type DocumentType = (typeof DocumentType)[keyof typeof DocumentType]


export const InviteStatus = {
  PENDING: 'PENDING',
  ACCEPTED: 'ACCEPTED',
  EXPIRED: 'EXPIRED'
} as const

export type InviteStatus = (typeof InviteStatus)[keyof typeof InviteStatus]


export const IncidentStatus = {
  OPEN: 'OPEN',
  INVESTIGATING: 'INVESTIGATING',
  MITIGATED: 'MITIGATED',
  RESOLVED: 'RESOLVED',
  FAILED: 'FAILED'
} as const

export type IncidentStatus = (typeof IncidentStatus)[keyof typeof IncidentStatus]


export const Severity = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
  CRITICAL: 'CRITICAL'
} as const

export type Severity = (typeof Severity)[keyof typeof Severity]


export const Level = {
  INFO: 'INFO',
  WARN: 'WARN',
  ERROR: 'ERROR',
  FATAL: 'FATAL',
  DEBUG: 'DEBUG'
} as const

export type Level = (typeof Level)[keyof typeof Level]


export const ActionStatus = {
  PENDING_APPROVAL: 'PENDING_APPROVAL',
  EXECUTING: 'EXECUTING',
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED',
  REJECTED: 'REJECTED'
} as const

export type ActionStatus = (typeof ActionStatus)[keyof typeof ActionStatus]